feature/image-upload used cloudinary for image upload. Changed format of posts. Changed the edit post functionality so text content can also be editted
