+++ controllers/api/post-routes.js
@@ -3,15 +3,13 @@ const sequelize = require('../../config/connection');
 const fs = require('fs');
 const { Post, User, Comment, Vote } = require('../../models');
 const withAuth = require('../../utils/auth')
-const multer = require('multer')
-const upload = multer({ dest: 'uploads/' })
 
 // get all users
 router.get('/', (req, res) => {
   Post.findAll({
     attributes: [
       'id',
-      'post_url',
+      'post_text_content',
       'title',
       'created_at',
       [sequelize.literal('(SELECT COUNT(*) FROM vote WHERE post.id = vote.post_id)'), 'vote_count']
@@ -45,7 +43,7 @@ router.get('/:id', (req, res) => {
     },
     attributes: [
       'id',
-      'post_url',
+      'post_text_content',
       'title',
       'created_at',
       [sequelize.literal('(SELECT COUNT(*) FROM vote WHERE post.id = vote.post_id)'), 'vote_count']
@@ -79,15 +77,15 @@ router.get('/:id', (req, res) => {
 });
 
 router.post('/', withAuth, (req, res, next) => {
-  // expects {title: 'Taskmaster goes public!', post_url: 'https://taskmaster.com/press', user_id: 1}
+  // expects {title: 'Taskmaster goes public!', post_text_content: 'https://taskmaster.com/press', user_id: 1}
   if (req.session) {
     Post.create({
       title: req.body.title,
-      post_url: req.body.post_url,
+      post_text_content: req.body.post_text_content,
       user_id: req.session.user_id,
       image_public_id: req.body.imagePublicID
     })
-      // .then(dbPostData => res.json(dbPostData))
+      .then(dbPostData => res.json(dbPostData))
       .catch(err => {
         console.log(err);
         res.status(500).json(err);
@@ -100,7 +98,7 @@ router.put('/upvote', withAuth, (req, res) => {
   if (req.session) {
     Post.upvote({ ...req.body, user_id: req.session.user_id }, { Vote, Comment, User })
       .then(updatedVoteData => res.json(updatedVoteData))
-      // .then(res.redirect('/dashboard'))
+      .then(res.redirect('/dashboard'))
       .catch(err => {
         console.log(err);
         res.status(500).json(err);
@@ -111,7 +109,8 @@ router.put('/upvote', withAuth, (req, res) => {
 router.put('/:id', withAuth, (req, res) => {
   Post.update(
     {
-      title: req.body.title
+      title: req.body.title,
+      post_text_content: req.body.post_text_content
     },
     {
       where: {
