+++ controllers/dashboard-routes.js
@@ -12,7 +12,7 @@ router.get('/', withAuth, (req, res) => {
       },
       attributes: [
         'id',
-        'post_url',
+        'post_text_content',
         'title',
         'created_at',
         'image_public_id',
@@ -36,7 +36,6 @@ router.get('/', withAuth, (req, res) => {
     .then(dbPostData => {
     // serialize data before passing to template
     const posts = dbPostData.map(post => post.get({ plain: true }));
-    (console.log('----------------------', posts))
     res.render('dashboard', { posts, loggedIn: true });
     })
     .catch(err => {
@@ -52,7 +51,7 @@ router.get('/edit/:id', withAuth, (req, res) => {
         },
         attributes: [
           'id',
-          'post_url',
+          'post_text_content',
           'title',
           'created_at',
           [sequelize.literal('(SELECT COUNT(*) FROM vote WHERE post.id = vote.post_id)'), 'vote_count']
