--- controllers/homeroutes.js
@@ -1,8 +1,0 @@
-const router = require('express').Router();
-
-router.get('/', (req, res) => {
-    res.render('homepage')
-
-});
-
-module.exports = router;
