+++ public/javascript/edit-post.js
@@ -2,12 +2,14 @@ async function editFormHandler(event) {
     event.preventDefault();
   
     const post_id = window.location.href.split('/')[5]
-    const post_title = document.querySelector('input[name="post-title"]').value.trim();
+    const postTitle = document.querySelector('input[name="post-title"]').value.trim();
+    const postTextContent = document.querySelector('textarea[name="post-text-content"]').value.trim();
 
     const response = await fetch(`/api/posts/${post_id}`, {
         method: 'PUT',
         body: JSON.stringify({
-            title: post_title
+            title: postTitle,
+            post_text_content: postTextContent
         }),
         headers: {
             'Content-Type': 'application/json'
