+++ utils/helpers.js
@@ -11,12 +11,4 @@ module.exports = {
 
         return word;
     },
-    format_url: url => {
-        return url
-            .replace('http://', '')
-            .replace('https://', '')
-            .replace('www.', '')
-            .split('/')[0]
-            .split('?'[0])[0];
-    },
 }
