+++ views/dashboard.handlebars
@@ -1,69 +1,33 @@
 <section>
     <h2>Create New Post</h2>
-    {{!-- <form action="/api/posts" method="POST" enctype="multipart/form-data" class="new-post-form">
-        <div>
-            <label for="post-title">Title</label>
-            <input type="text" id="post_title" name="post_title" />
-        </div>
-        <div>
-            <label for="post-url">Link</label>
-            <input type="text" id="post_url" name="post_url" />
-        </div>
-        <div>
-            <label for="image-upload">Upload An Image</label>
-            <input type="file" id="image" name="image">
-        </div>
-        <button type="submit" class="btn">Create</button>
-        <button id="upload_widget" class="cloudinary-button">Upload Image</button>
-    </form> --}}
     <form class="new-post-form">
         <div>
             <label for="post_title">Title</label>
             <input type="text" id="post_title" name="post_title" />
         </div>
         <div>
-            <label for="post_url">Link</label>
-            <input type="text" id="post_url" name="post_url" />
-        </div>
-        <div>
-            <label for="image-upload">Upload An Image</label>
-            <input type="file" id="image" name="image">
+            <label for="post_text_content">Say Something About Your Dog</label>
+            <input type="text" id="post_text_content" name="post_text_content" />
         </div>
         <button type="submit" class="btn">Create</button>
-        <button id="upload_widget" class="cloudinary-button">Upload Image</button>
+        <button id="upload_widget" class="btn">Attach Image</button>
     </form>
     <div class="preview-images"></div>
 </section>
 
 {{#if posts.length}}
 <section>
-    <h2>Your Posts</h2>
-    <ol>
+    <h2 class="d-flex justify-content-center border-bottom border-primary mb-4 mt-4">Your Posts</h2>
+    <ul class="post-list">
         {{#each posts as |post|}}
-        <li>
+        <li class = "border-bottom border-secondary mb-4 bg-secondary rounded">
             {{> post-info post}}
-            <a href="/dashboard/edit/{{post.id}}" class="edit-link">Edit post</a>
+            <a href="/dashboard/edit/{{post.id}}" class="edit-link d-flex justify-content-center">Edit post</a>
         </li>
         {{/each}}
-    </ol>
+    </ul>
 </section>
 {{/if}}
 
-{{!-- <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script> --}}
 <script src="https://upload-widget.cloudinary.com/global/all.js" type="text/javascript"></script>  
-<script src="/javascript/cloudinary.js"></script>
-
-<script type="module" src="/javascript/cloudinary.js">  
-var myWidget = cloudinary.createUploadWidget({
-  cloudName: 'dr3wpa5jd', 
-  uploadPreset: 'xie2ddac'}, (error, result) => { 
-    if (!error && result && result.event === "success") { 
-      console.log('Done! Here is the image info: ', result.info); 
-    }
-  }
-)
-
-document.getElementById("upload_widget").addEventListener("click", function(){
-    myWidget.open();
-  }, false);
-</script>
+<script src="/javascript/add-post.js"></script>
