+++ views/edit-post.handlebars
@@ -5,8 +5,10 @@
   </h2>
   <form class="edit-post-form">
     <div>
-      <input name="post-title" type="text" value="{{post.title}}" />
-      <span>({{format_url post.post_url}})</span>
+      <label for="post-title">Edit the title of the post</label>
+      <input id="post-title" name="post-title" type="text" value="{{post.title}}" class="mb-4"/>
+      <label for="text-content">Edit the text content of the post</label>
+      <textarea id='text-content' name="post-text-content">{{post.post_text_content}}</textarea>
     </div>
     <div>
       {{post.vote_count}} {{format_plural "point" post.vote_count}} by you on {{format_date post.created_at}}
