+++ views/homepage.handlebars
@@ -1,13 +1,8 @@
-{{!-- <ol class="post-list">
-  {{#each posts as |post|}} --}}
-  {{!-- <li> --}}
-    <ol class="post-list">
-      {{#each posts as |post|}}
-      <li>
-        {{> post-info post }}
-      </li>
-      {{/each}}
-    </ol>
-  {{!-- </li> --}}
-  {{!-- {{/each}}
-</ol> --}}
+
+<ol class="post-list">
+  {{#each posts as |post|}}
+  <li>
+    {{> post-info post }}
+  </li>
+  {{/each}}
+</ol>
